+++ packages/components/src/index.native.js
@@ -89,6 +89,7 @@ export { default as ImageEditingButton } from './mobile/image/image-editing-butt
 export { default as InserterButton } from './mobile/inserter-button';
 export { setClipboard, getClipboard } from './mobile/clipboard';
 export { default as AudioPlayer } from './mobile/audio-player';
+export { default as Badge } from './mobile/badge';
 
 // Utils
 export { colorsUtils } from './mobile/color-settings/utils';